requirements.txt
pandas
torch
transformers
pillow
pyodbc
sqlalchemy
ydata_profiling